Added Max Line Limit, Updated Change Log

+++ src/main/java/drawingbot/api/IPlottingTask.java
@@ -17,7 +17,7 @@ public interface IPlottingTask {
      * @param progress the current progress
      * @param max the max progress
      */
-    void updateProgess(float progress, float max);
+    void updateProgess(double progress, double max);
 
     /**
      * Notifies the PlottingTask that the {@link IPathFindingModule} has finished
